requirements.txt
streamlit
langchain
langchain_openai
langchain_community
pdfminer.six
python-docx